+++ database/install/myems_user_db.sql
@@ -8,6 +8,20 @@ CREATE DATABASE IF NOT EXISTS `myems_user_db` CHARACTER SET 'utf8mb4' COLLATE 'u
 USE `myems_user_db` ;
 
 -- ---------------------------------------------------------------------------------------------------------------------
+-- Table structure for `myems_user_db`.tbl_api_keys
+-- ---------------------------------------------------------------------------------------------------------------------
+DROP TABLE IF EXISTS `myems_user_db`.`tbl_api_keys`;
+
+CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_api_keys`  (
+  `id` int NOT NULL AUTO_INCREMENT,
+  `name` VARCHAR(128) NOT NULL,
+  `token` VARCHAR(255) NOT NULL,
+  `created_datetime_utc` DATETIME NOT NULL,
+  `expires_datetime_utc` DATETIME NOT NULL,
+  PRIMARY KEY (`id`));
+CREATE INDEX `tbl_api_keys_index_1` ON `myems_user_db`.`tbl_api_keys` (`created_datetime_utc`, `name`);
+
+-- ---------------------------------------------------------------------------------------------------------------------
 -- Table `myems_user_db`.`tbl_email_messages`
 -- ---------------------------------------------------------------------------------------------------------------------
 DROP TABLE IF EXISTS `myems_user_db`.`tbl_email_messages`;
@@ -142,9 +156,9 @@ CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_notifications` (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_notifications_index_1` ON  `myems_user_db`.`tbl_notifications`  (`user_id`, `created_datetime_utc`, `status`);
 
--- ----------------------------
+-- ---------------------------------------------------------------------------------------------------------------------
 -- Table structure for `myems_user_db`.tbl_verification_codes
--- ----------------------------
+-- ---------------------------------------------------------------------------------------------------------------------
 DROP TABLE IF EXISTS `myems_user_db`.`tbl_verification_codes`;
 
 CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_verification_codes`  (
@@ -155,17 +169,3 @@ CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_verification_codes`  (
   `expires_datetime_utc` DATETIME NOT NULL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_verirication_codes_index_1` ON `myems_user_db`.`tbl_verification_codes` (`recipient_email`, `created_datetime_utc`);
-
--- ----------------------------
--- Table structure for `myems_user_db`.tbl_api_keys
--- ----------------------------
-DROP TABLE IF EXISTS `myems_user_db`.`tbl_api_keys`;
-
-CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_api_keys`  (
-  `id` int NOT NULL AUTO_INCREMENT,
-  `name` VARCHAR(128) NOT NULL,
-  `token` VARCHAR(255) NOT NULL,
-  `created_datetime_utc` DATETIME NOT NULL,
-  `expires_datetime_utc` DATETIME NOT NULL,
-  PRIMARY KEY (`id`));
-CREATE INDEX `tbl_api_keys_index_1` ON `myems_user_db`.`tbl_api_keys` (`created_datetime_utc`, `name`);