+++ database/upgrade/upgrade3.6.0RC.sql
@@ -44,11 +44,6 @@ CREATE TABLE IF NOT EXISTS `myems_system_db`.`tbl_microgrids_windturbines` (
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_microgrids_windturbines_index_1` ON  `myems_system_db`.`tbl_microgrids_windturbines` (`name`);
 
--- ----------------------------
--- Table structure for `myems_user_db`.tbl_verification_codes
--- ----------------------------
-DROP TABLE IF EXISTS `myems_user_db`.`tbl_verification_codes`;
-
 CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_verification_codes`  (
   `id` int NOT NULL AUTO_INCREMENT,
   `recipient_email` VARCHAR(128) NOT NULL,
@@ -57,11 +52,6 @@ CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_verification_codes`  (
   `expires_datetime_utc` DATETIME NOT NULL,
   PRIMARY KEY (`id`));
 CREATE INDEX `tbl_verirication_codes_index_1` ON `myems_user_db`.`tbl_verification_codes` (`recipient_email`, `created_datetime_utc`);
-
--- ----------------------------
--- Table structure for `myems_user_db`.tbl_api_keys
--- ----------------------------
-DROP TABLE IF EXISTS `myems_user_db`.`tbl_api_keys`;
 
 CREATE TABLE IF NOT EXISTS `myems_user_db`.`tbl_api_keys`  (
   `id` int NOT NULL AUTO_INCREMENT,