+++ myems-admin/app/controllers/users/apikey/apikey.controller.js
@@ -9,7 +9,8 @@ app.controller('ApiKeyController', function (
 	SweetAlert) {
 	$scope.cur_user = JSON.parse($window.localStorage.getItem("myems_admin_ui_current_user"));
 	$scope.getAllApiKeys = function () {
-		ApiKeyService.getAllApiKeys(function (response) {
+		let headers = { "User-UUID": $scope.cur_user.uuid, "Token": $scope.cur_user.token };
+		ApiKeyService.getAllApiKeys(headers, function (response) {
 			if (angular.isDefined(response.status) && response.status === 200) {
 				$scope.apiKeys = response.data;
 			} else {