+++ CHANGELOG.md
@@ -20,7 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - added access_contorl to shopfloor reports actions in myems-api
 - added svg, inverters and converters to microgrid in system database
 - added svg to microgrid in myems-api and myems-admin
-
+- added API Key to database, myems-api and myems-admin
+- 
 ### Changed
 - renamed access_control to admin_control in myems-api
 - changed forgot password and user register processes